fix: ApplicaitonData unkwnown pareses all to 'extra'
For now, until we can build out better logic around parsing
Group / Variation fields, if we encounter one that is now a part
of our set of "known", we can just assign all the (remaining)
applicaiton data to the 'extra' field. This allows for
FromBytes -> ToBytes to return equivelent data, but its just 'data'
and we don't actually parse it to understand the Group/Variation
data objects etc.