build: CI Actions (lint/test, release), hooks
* linting using golangci-lint now applied at pre-commit
hook and via GitHub Actions for any push (i.e. PR). Most
of the changes in this should just be formatting and
making the linter happy. In the future we will want to go
back and start adding back some of the checks and improving
the quality of code to make them happy, but for now this is
better than it was.
* pre-commit hook is provided so that contributors can ensure
that both linting and testing pass before a commit is
generated. This needs to be configured for each developer,
just run 'make hooks'. Installing golangci-lint is left to the
user.
* makefile is provided to lint, fix, test
* CI actions for auto lint/test and release on PR approval.
This is my first time configuring the auto-release so we might
have to come back and make some changes